requirements.txt
aiogram==3.7.0
aiohttp>=3.8.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
pandas>=2.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
plotly>=5.17.0
python-telegram-bot~=22.5